Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/url-with-empty-query-expected.txt

Make sure handshake with URL with empty query components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/websocket/tests/hybi/echo-path?
WebSocket is open
received:/websocket/tests/hybi/echo-path?
WebSocket is closed
PASS handshake_success is true
PASS path is expected_path
PASS successfullyParsed is true

TEST COMPLETE
